// asr.webapp.v2/src/app/modules/inventory/transfer/transfer.types.ts
import { Paging } from "app/core/type/paging/paging.type";

export class Transfer {
    transferId: number;
    transferCode: string;
    referenceNo: string;
    transferTypeId: number;
    transferTypeName: string;
    merchandiseCode: string;
    merchandiseName: string;
    merchandiseTypeId: number;
    merchandiseTypeName: string;
    srcOwnerId: number;
    srcOwnerName: string;
    srcOwnerTypeId: number;
    srcOwnerTypeName: string;
    srcWarehouseId: number;
    srcWarehouseName: string;
    srcWarehouseTypeId: number;
    destnOwnerId: number;
    destnOwnerName: string;
    destnOwnerTypeId: number;
    destnOwnerTypeName: string;
    destnWarehouseId: number;
    destnWarehouseName: string;
    destnWarehouseTypeId: number;
    orderId: number;
    transferDetail: Array<TransferDetail>;
    merchandiseType: number;
    measure: number;
    description: string;
    status: number;
    statusDescription: string;
    statusColor: string;
    languageID: string;
    clientIP: string;
    registerUserID: number;
    registerUserName: string;
    localChangeDate: string;
    localChangeTime: string;
    chaneDate: string;
    requestTypeId: number;
    requestTypeName: string;
    statusIdList: Array<number>;
    srcOwnerIdList: Array<number>;
    srcWarehouseIdList: Array<number>;
    destnOwnerIdList: Array<number>;
    destnWarehouseIdList: Array<number>;
    transferTypeIdList: Array<number>;
    requestTypeIdList: Array<number>;
    page: Paging;
}

export class TransferDetail {
    transferId: number;
    transferDetailId: number;
    transferCode: string;
    transferTypeId: number;
    transferTypeName: string;
    srcOwnerId: number;
    srcOwnerName: number;
    srcOwnerTypeId: number;
    serOwnerTypeName: string;
    srcWarehouseId: number;
    srcWarehouseName: number;
    destnOwnerId: number;
    destnOwnerName: string;
    destnOwnerTypeId: number;
    destnOwnerTypeName: string;
    destnWarehouseName: string;
    destnWarehouseTypeId: number;
    merchandiseId: number;
    merchandiseCode: string;
    merchandiseName: string;
    merchandiseTypeId: number;
    merchandiseTypeName: string;
    srcCoverId: number;
    srcCoverName: string;
    destnCoverId: number;
    destnCoverName: string;
    merchandiseMeasure: number;
    measure: number;
    unitId: number;
    unitName: string;
    serial: string;
    status: number;
    statusDescription: string;
    statusColor: string;
    requestTypeId: number;
    requestId: number;
    requestDetailId: number;
    statusIdList: Array<number>;
    srcOwnerIdList: Array<number>;
    srcWarehouseIdList: Array<number>;
    destnOwnerIdList: Array<number>;
    destnWarehouseIdList: Array<number>;
    transferTypeIdList: Array<number>;
    languageID: string;
    clientIP: string;
    registerUserID: number;
    registerUserName: string;
    localChangeDate: string;
    localChangeTime: string;
    chaneDate: string;
    page: Paging;

}